google-patents-bq-test/requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
requests
openai
google-generativeai
anthropic
